requirements.txt
tortoise_orm~=0.19.1
fastapi==0.79.0
pydantic~=1.9.1
starlette~=0.19.1
uvicorn~=0.18.2
aioredis~=2.0.1
python-jose~=3.3.0
asyncpg~=0.25.0
requests~=2.24.0
aiosmtplib~=1.1.6
python-multipart~=0.0.5
python-consul~=1.1.0